corrected the comments with the review
