+++ front/src/components/Header/styles.module.scss
@@ -12,7 +12,10 @@
 
 .header {
   font-weight: 600;
-  cursor: pointer;
+}
+
+.headerLink {
+  text-decoration: none;
 }
 
 .subHeader {
