+++ front/src/redux/Area.js
@@ -45,27 +45,27 @@ export default function area(state = defaultArea, { type, payload }) {
   let areaId = null;
   switch (type) {
     case SET_AREA_ACTION:
-      return { ...state, ...{ countries: payload } };
+      return { ...state, countries: payload };
     case SET_COUNTRY_ACTION:
-      const country = state.countries.filter((item) => {
+      const country = state.countries.find((item) => {
         return item.id == payload;
-      }).pop();
+      });
       const regions = country.areas;
       if (regions.length == 0) {
         areaId = payload;
       }
-      return { ...state, ...{ countryId: payload, regionId: '', cityId: '', regions: regions, cities: [], areaId: areaId } };
+      return { ...state, countryId: payload, regionId: '', cityId: '', regions: regions, cities: [], areaId: areaId };
     case SET_REGION_ACTION:
-      const region = state.regions.filter((item) => {
+      const region = state.regions.find((item) => {
         return item.id == payload
-      }).pop();
+      });
       const cities = region.areas;
       if (cities.length == 0) {
         areaId = payload;
       }
-      return { ...state, ...{ regionId: payload, cityId: '', cities: cities, areaId: areaId } };
+      return { ...state, regionId: payload, cityId: '', cities: cities, areaId: areaId };
     case SET_CITY_ACTION:
-      return { ...state, ...{ cityId: payload, areaId: payload } };
+      return { ...state, cityId: payload, areaId: payload };
     default:
       return state;
   }
